pyproject.toml
[project]
name = "sotrans-analytics-department"
version = "0.0.0"
description = "Проекты аналитического отдела ГК \"Сотранс\""
authors = [
    {name = "Dmitry 'ts.kdv.raz0r' Kolganov",email = "ts.kdv.raz0r@ya.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pywin32 (>=310,<311)",
    "pandas (>=2.2.3,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
